Change to default 'formatoptions'
Added 'a' and 'c' flags, which makes vim automatically wrap
comments when inserting, and automatically re-wrap when new
text is inserted (like gqap after every keystroke). Downside
is when you want precise formatting of your comments (e.g.,
for folding). Habit to get into is ':set fo-=a' when this
becomes annoying.